# Annotator panel optimization