--- a/grails-app/domain/org/bbop/apollo/FeatureLocation.groovy
+++ b/grails-app/domain/org/bbop/apollo/FeatureLocation.groovy
@@ -1,7 +1,9 @@
 package org.bbop.apollo
 
 class FeatureLocation {
-
+    static mapping = {
+        length formula: 'FMAX-FMIN'
+    }
     static constraints = {
         feature nullable: false
         fmin nullable: false
@@ -19,16 +21,17 @@ class FeatureLocation {
 
     }
 
-    Feature feature;
-    Integer fmin;
-    boolean isFminPartial;
-    Integer fmax;
-    boolean isFmaxPartial;
-    Integer strand;
-    Integer phase;
-    String residueInfo;
-    int locgroup;
-    int rank;
+    Feature feature
+    Integer fmin
+    Integer length
+    boolean isFminPartial
+    Integer fmax
+    boolean isFmaxPartial
+    Integer strand
+    Integer phase
+    String residueInfo
+    int locgroup
+    int rank
     Sequence sequence
 
     static belongsTo = [Feature]
@@ -63,6 +66,9 @@ class FeatureLocation {
         return fmax-fmin
     }
 
+    public Integer getFeatureLength(){
+        return fmax-fmin
+    }
     public FeatureLocation generateClone() {
         FeatureLocation cloned = new FeatureLocation();
         cloned.sequence = this.sequence;