--- a/grails-app/domain/org/bbop/apollo/FeatureLocation.groovy
+++ b/grails-app/domain/org/bbop/apollo/FeatureLocation.groovy
@@ -9,8 +9,7 @@ class FeatureLocation {
         fmin nullable: false
         fmax nullable: false
         sequence nullable: false
-
-
+        length nullable: true
         isFminPartial nullable: true
         isFmaxPartial nullable: true
         strand nullable: true
@@ -18,7 +17,6 @@ class FeatureLocation {
         residueInfo nullable: true
         locgroup nullable: true
         rank nullable: true
-
     }
 
     Feature feature