This is a special query that optimizes the annotator panel for #820

It does a two step query

1. Get the genes in a specific range for the sidebar
2. Get the details about the genes in that specific range

This blog nicely summarizes that issue http://floledermann.blogspot.com/2007/10/solving-hibernate-criterias-distinct.html (briefly, the first step doesn't use select distinct() and the second one does)

We also have a "convertFeatureToJSONLite" for the purpose that avoids getting extraneous feature info, but even if we did want to display that info, we would still do well by using this strategy

I get about ~300ms for selecting new pages (mean 311ms, stddev 14ms) with database that is bulkloaded with over 10000 genes

Same database on  current master (mean 4528ms, stddev 775ms), so it's quite improved