@nathandunn 2016-02-02
4 seconds to 150 ms . . nice for 1800 features.  

The "length" sorting is not working, however.   Also, it should always default from largest to smallest.