@cmdcolin 2016-02-02
Ah. I had commented that out awhile ago. I am not sure how to do that in the criteria query. Can we just disable sorting by feature length? I am not sure what purpose it serves