@nathandunn 2016-02-02
I was told that it was very common to sort by length starting large, so let's leave it in. 

In the old code where FeatureLocation is fl 

```
abs(fl.fmax-fl.fmin) as seqLength
order by seqLength
```

So will probably be something like this in the new code:

```
     if(sort=="length") {
       order('abs(fmin-fmax)', sortorder)
 }
```

Or something like that.  You might need an alias, don't know.   I don't think it will be hard.