@cmdcolin 2016-02-02
I am going off of this and saw that I can make a "derived" property that isn't persisted in database  using a sql like formula http://grails.github.io/grails-doc/2.4.5/guide/GORM.html#derivedProperties


Seems to solve the prolbem.

 I also realized that I had to also add sorting on the "second" query too because otherwise that one can come in out-of-order

Anyways, should be ready now.