@cmdcolin 2016-02-02
I guess I have to add "length nullable: true" to the constraints. seems ok now. I checked and it is still not a column in the database, so weird that it needs to be allowed nullable, but seems ok now