@nathandunn 2016-02-02
You don't want to add a property to FeatureLocation.  Also, not sure how this is populated.   

I think what you want is a projection most likely . . and more concretely a sql projection.  You could also use a detachedcriteria query to do a subselect, but I think the sql projection will be easier.

http://grails.github.io/grails-doc/2.4.x/guide/GORM.html (just search for it)
https://docs.jboss.org/hibernate/orm/3.5/api/org/hibernate/criterion/Projections.html
http://www.devarticles.com/c/a/Java/Hibernate-Criteria-Queries-in-Depth/1/