@nathandunn 2016-02-02
"getLength()" is equivalent to "length" in groovy, but when doing a query you can't use a derived property in that  way (since it is actually directly hitting the database).  

Worst-case you can go back to HQL, but I think that the SQL Projection will work well.