@cmdcolin 2016-02-02
>but when doing a query you can't use a derived property in that way (since it is actually directly hitting the database).

Actually you can use the derived property, for sorting, which is what I did here