@nathandunn 2016-02-02
Oh I see . . missed this line  . . .  https://github.com/GMOD/Apollo/pull/843/files#diff-7ddee18834450388098b372e55b95e1bR4

That does the equivalent of what a SQL Projection would do.    I'll take a look.